addressed PR review comments
