+++ quickfixj-from-fix-orchestra/pom.xml
@@ -1,0 +1,27 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>org.quickfixj</groupId>
+		<artifactId>quickfixj-parent</artifactId>
+		<version>3.0.0-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>quickfixj-from-fix-orchestra-parent</artifactId>
+	<packaging>pom</packaging>
+
+	<name>quickfixj-from-fix-orchestra-parent</name>
+	<description>Parent for  QuickFIX/J from FIX Orchestra Repository tools</description>
+	<url>http://www.quickfixj.org</url>
+	
+	<properties>
+	</properties>
+
+
+	<modules>
+		<module>quickfixj-from-fix-orchestra</module>
+		<module>quickfixj-codegen-from-fix-orchestra-maven-plugin</module>
+		<module>quickfixj-dictionary-from-fix-orchestra-maven-plugin</module>
+	</modules>
+
+</project>
