+++ quickfixj-from-fix-orchestra/quickfixj-codegen-from-fix-orchestra-maven-plugin/src/it/simple-it/verify.groovy
