+++ quickfixj-from-fix-orchestra/quickfixj-dictionary-from-fix-orchestra-maven-plugin/readme.md
@@ -1,0 +1,5 @@
+# Repository to QuickFIX/J Dictionary Maven Plugin
+
+![](../FIXorchestraLogo.png)
+
+This module provides the Maven Plugin to generate a QuickFIX/J Dictionary file given a FIX Orchestra repository file.
