+++ quickfixj-from-fix-orchestra/quickfixj-dictionary-from-fix-orchestra-maven-plugin/src/it/simple-it/pom.xml
