+++ quickfixj-from-fix-orchestra/quickfixj-dictionary-from-fix-orchestra-maven-plugin/src/test/java/org/quickfixj/DictionaryGeneratorMojoTest.java
