+++ quickfixj-from-fix-orchestra/quickfixj-from-fix-orchestra/pom.xml
@@ -1,0 +1,81 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>org.quickfixj</groupId>
+		<artifactId>quickfixj-from-fix-orchestra-parent</artifactId>
+		<version>3.0.0-SNAPSHOT</version>
+	</parent>
+	
+	<artifactId>quickfixj-from-fix-orchestra</artifactId>
+	<name>quickfixj-from-fix-orchestra</name>
+	<description>Generates QuickFIX data dictionary and message classes from a FIX Orchestra file</description>
+	<url>http://www.quickfixj.org</url>
+	
+	<properties>
+		<picocli-version>4.6.3</picocli-version>
+	</properties>
+
+	<dependencies>
+		<dependency>
+			<groupId>io.fixprotocol.orchestra</groupId>
+			<artifactId>repository</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter-engine</artifactId>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter-params</artifactId>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+		    <groupId>commons-io</groupId>
+		    <artifactId>commons-io</artifactId>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>jakarta.xml.bind</groupId>
+			<artifactId>jakarta.xml.bind-api</artifactId>
+		</dependency>
+		<dependency>
+		  <groupId>info.picocli</groupId>
+		  <artifactId>picocli</artifactId>
+		  <version>${picocli-version}</version>
+		</dependency>		
+		<dependency>
+			<groupId>com.sun.xml.bind</groupId>
+			<artifactId>jaxb-impl</artifactId>
+			<scope>runtime</scope>
+		</dependency>
+	</dependencies>
+
+	<build>
+		<plugins>
+			<plugin>
+			  <groupId>org.apache.maven.plugins</groupId>
+			  <artifactId>maven-compiler-plugin</artifactId>
+			  <configuration>
+			    <annotationProcessorPaths>
+			      <path>
+			        <groupId>info.picocli</groupId>
+			        <artifactId>picocli-codegen</artifactId>
+			        <version>${picocli-version}</version>
+			      </path>
+			    </annotationProcessorPaths>
+			    <compilerArgs>
+			      <arg>-Aproject=${project.groupId}/${project.artifactId}</arg>
+			    </compilerArgs>
+			  </configuration>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-surefire-plugin</artifactId>
+			</plugin>
+		</plugins>
+	</build>
+
+</project>
