+++ quickfixj-from-fix-orchestra/quickfixj-from-fix-orchestra/readme.md
@@ -1,0 +1,11 @@
+# repository-to-quickfixj
+
+![](../FIXorchestraLogo.png)
+
+This module is derived from the FIX Trading Community [fix-orchestra-quickfix](https://github.com/FIXTradingCommunity/fix-orchestra-quickfix).
+Accordingly the Apache 2.0 license is applicable.
+
+This module provides both QuickFIX/J Dictionary and code generation given a FIX Orchestra repository file.
+
+This method of code generation is faster than the xslt transformations used to build from QuickFIX/J dictionary specifications.
+It was not possible to meet build time constraints using the XSLT approach when building the complete FIX Latest standard.
