+++ quickfixj-from-fix-orchestra/quickfixj-from-fix-orchestra/src/test/java/io/fixprotocol/orchestra/quickfix/DataDictionaryGeneratorTest.java
