--- repository-to-quickfixj-parent/pom.xml
@@ -1,27 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.quickfixj</groupId>
-		<artifactId>quickfixj-parent</artifactId>
-		<version>3.0.0-SNAPSHOT</version>
-	</parent>
-
-	<artifactId>repository-to-quickfixj-parent</artifactId>
-	<packaging>pom</packaging>
-
-	<name>repository-to-quickfixj-parent</name>
-	<description>Parent for FIX Orchestra Repository to QuickFIX/J tools</description>
-	<url>http://www.quickfixj.org</url>
-	
-	<properties>
-	</properties>
-
-
-	<modules>
-		<module>repository-to-quickfixj</module>
-		<module>repository-to-qfj-codegen-maven-plugin</module>
-		<module>repository-to-qfj-dictionary-maven-plugin</module>
-	</modules>
-
-</project>
