--- repository-to-quickfixj-parent/repository-to-qfj-dictionary-maven-plugin/src/test/resources/project-to-test/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>org.quickfixj</groupId>
-	<artifactId>project-to-test</artifactId>
-	<version>1.0-SNAPSHOT</version>
-	<packaging>jar</packaging>
-	<name>Test MyMojo</name>
-
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.quickfixj</groupId>
-				<artifactId>repository-to-qfj-dictionary-maven-plugin</artifactId>
-				<configuration>
-					<orchestration>resources/trade-latest.xml</orchestration>
-					<outputDirectory>target/generated-resources</outputDirectory>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
