--- repository-to-quickfixj-parent/repository-to-quickfixj/pom.xml
@@ -1,99 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>org.quickfixj</groupId>
-		<artifactId>repository-to-quickfixj-parent</artifactId>
-		<version>3.0.0-SNAPSHOT</version>
-	</parent>
-	
-	<artifactId>repository-to-quickfixj</artifactId>
-	<name>repository-to-quickfixj</name>
-	<description>Generates QuickFIX data dictionary and message classes from a FIX Orchestra file</description>
-	<url>http://www.quickfixj.org</url>
-	
-	<properties>
-		<picocli-version>4.6.3</picocli-version>
-	</properties>
-
-	<dependencies>
-		<dependency>
-			<groupId>io.fixprotocol.orchestra</groupId>
-			<artifactId>repository</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter-engine</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter-params</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-		    <groupId>commons-io</groupId>
-		    <artifactId>commons-io</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>jakarta.xml.bind</groupId>
-			<artifactId>jakarta.xml.bind-api</artifactId>
-		</dependency>
-		<dependency>
-		  <groupId>info.picocli</groupId>
-		  <artifactId>picocli</artifactId>
-		  <version>${picocli-version}</version>
-		</dependency>		
-		<dependency>
-			<groupId>com.sun.xml.bind</groupId>
-			<artifactId>jaxb-impl</artifactId>
-			<scope>runtime</scope>
-		</dependency>
-	</dependencies>
-
-	<build>
-		<plugins>
-			<plugin>
-			  <groupId>org.apache.maven.plugins</groupId>
-			  <artifactId>maven-compiler-plugin</artifactId>
-			  <configuration>
-			    <annotationProcessorPaths>
-			      <path>
-			        <groupId>info.picocli</groupId>
-			        <artifactId>picocli-codegen</artifactId>
-			        <version>${picocli-version}</version>
-			      </path>
-			    </annotationProcessorPaths>
-			    <compilerArgs>
-			      <arg>-Aproject=${project.groupId}/${project.artifactId}</arg>
-			    </compilerArgs>
-			  </configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-			</plugin>
-			<plugin>
-				<groupId>org.jboss.tattletale</groupId>
-				<artifactId>tattletale-maven</artifactId>
-				<executions>
-					<execution>
-						<phase>verify</phase>
-						<goals>
-							<goal>report</goal>
-						</goals>
-					</execution>
-				</executions>
-				<configuration>
-					<!-- This is the location which will be scanned for generating tattletale reports -->
-					<source>${project.build.directory}</source>
-					<!-- This is where the reports will be generated -->
-					<destination>${project.build.directory}/tattle-tale-report</destination>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-
-</project>
