--- repository-to-quickfixj-parent/repository-to-quickfixj/readme.md
@@ -1,11 +1,0 @@
-# repository-to-quickfixj
-
-![](../FIXorchestraLogo.png)
-
-This module is derived from the FIX Trading Community [fix-orchestra-quickfix](https://github.com/FIXTradingCommunity/fix-orchestra-quickfix).
-Accordingly the Apache 2.0 license is applicable.
-
-This module provides code generation and QuickFIX/J message dictionary generation for QuickFIX/J from a FIX Orchestration.
-
-This method of code generation is faster than the xslt transformations used to build from QuickFIX/J dictionary specifications.
-It was not possible to meet build time constraints using the xslt approach when building from the complete FIX standard orchestration.
